Move the existing test script to a file called demo-spec.js. Create a proper wp-log-in-out-spec.js test file that will contain migrated tests.
